requirements.txt
# Django framework requirements
Django==3.0.10
djangorestframework==3.11.1
tensorflow==2.3.0
mod_wsgi
